Add mc1.21.8 support
