--- magiclib-better-dev/versions/1.21.7-fabric/gradle.properties
@@ -1,20 +1,0 @@
-# Dependency Versions
-dependencies.minecraft_dependency=>=1.21.6- <1.21.8-
-dependencies.minecraft_version=1.21.7
-
-# Fabric API 0.128.2+1.21.7
-dependencies.api.fabric_version=0.128.2+1.21.7
-# Mod Menu 15.0.0-beta.3
-# modmenu-15.0.0-beta.3.jar
-dependencies.api.modmenu_version=15.0.0-beta.3
-
-# IMBlockerFabric 1.0.24
-# imblockerfabric-1.0.24.jar
-dependencies.runtime.imblocker_version=1.0.24
-# In-Game Account Switcher 9.0.4
-# IAS-Fabric-1.21.7-9.0.4.jar
-dependencies.runtime.inGameAccountSwitcher_version=Fs2YTzMh
-
-# Publish properties
-publish.game_version=1.21.6,1.21.7
-publish.dependencies_list=
