--- magiclib-better-dev/versions/1.21.7-neoforge/gradle.properties
@@ -1,18 +1,0 @@
-# Dependency Versions
-dependencies.neoforge_version=21.7.2-beta
-dependencies.minecraft_dependency=1.21.7
-dependencies.minecraft_version=1.21.7
-
-# IMBlocker 5.3.1
-# IMBlocker-5.3.1-neoforge-1.20.6+.jar
-dependencies.runtime.imblocker_version=CIdH0JQr
-# In-Game Account Switcher 9.0.4
-# IAS-NeoForge-1.21.7-9.0.4.jar
-dependencies.runtime.inGameAccountSwitcher_version=bxPb5zqK
-
-# Loom Properties
-loom.platform=neoforge
-
-# Publish properties
-publish.game_version=1.21.7
-publish.dependencies_list=
