+++ magiclib-better-dev/versions/1.21.8-fabric/gradle.properties
@@ -1,0 +1,20 @@
+# Dependency Versions
+dependencies.minecraft_dependency=>=1.21.6- <1.21.9-
+dependencies.minecraft_version=1.21.8
+
+# Fabric API 0.129.0+1.21.8
+dependencies.api.fabric_version=0.129.0+1.21.8
+# Mod Menu 15.0.0-beta.3
+# modmenu-15.0.0-beta.3.jar
+dependencies.api.modmenu_version=15.0.0-beta.3
+
+# IMBlockerFabric 1.0.24
+# imblockerfabric-1.0.24.jar
+dependencies.runtime.imblocker_version=1.0.24
+# In-Game Account Switcher 9.0.5
+# IAS-Fabric-1.21.8-9.0.5.jar
+dependencies.runtime.inGameAccountSwitcher_version=9.0.5
+
+# Publish properties
+publish.game_version=1.21.6,1.21.7,1.21.8
+publish.dependencies_list=
