+++ magiclib-better-dev/versions/1.21.8-fabric/src/main/resources/magiclib-better-dev.accesswidener
