+++ magiclib-better-dev/versions/1.21.8-neoforge/gradle.properties
@@ -1,0 +1,18 @@
+# Dependency Versions
+dependencies.neoforge_version=21.8.5-beta
+dependencies.minecraft_dependency=1.21.8
+dependencies.minecraft_version=1.21.8
+
+# IMBlocker 5.4.0
+# IMBlocker-5.4.0-neoforge-1.20.6+.jar
+dependencies.runtime.imblocker_version=CouZoFOg
+# In-Game Account Switcher 9.0.5
+# IAS-NeoForge-1.21.8-9.0.5.jar
+dependencies.runtime.inGameAccountSwitcher_version=eQRA4vob
+
+# Loom Properties
+loom.platform=neoforge
+
+# Publish properties
+publish.game_version=1.21.6,1.21.7,1.21.8
+publish.dependencies_list=
