+++ magiclib-better-dev/versions/1.21.8-neoforge/src/main/resources/magiclib-better-dev.accesswidener
