+++ magiclib-better-dev/versions/mapping-1.21.8-fabric-neoforge.txt
