+++ magiclib-legacy-compat/build.gradle
@@ -29,7 +29,7 @@ preprocess {
     Node mc_12103_fabric = createNode("legacy-1.21.3-fabric", 1_21_03, "mojang")
     Node mc_12104_fabric = createNode("legacy-1.21.4-fabric", 1_21_04, "mojang")
     Node mc_12105_fabric = createNode("legacy-1.21.5-fabric", 1_21_05, "mojang")
-    Node mc_12107_fabric = createNode("legacy-1.21.7-fabric", 1_21_07, "mojang")
+    Node mc_12108_fabric = createNode("legacy-1.21.8-fabric", 1_21_08, "mojang")
 
     mc_11502_fabric.link(mc_11404_fabric, file("versions/mapping-1.15.2-1.14.4.txt"))
     mc_11605_fabric.link(mc_11502_fabric, file("versions/mapping-1.16.5-1.15.2.txt"))
@@ -46,5 +46,5 @@ preprocess {
     mc_12101_fabric.link(mc_12103_fabric, file("versions/mapping-1.21.1-1.21.3.txt"))
     mc_12103_fabric.link(mc_12104_fabric, file("versions/mapping-1.21.3-1.21.4.txt"))
     mc_12104_fabric.link(mc_12105_fabric, file("versions/mapping-1.21.4-1.21.5.txt"))
-    mc_12105_fabric.link(mc_12107_fabric, file("versions/mapping-1.21.5-1.21.7.txt"))
+    mc_12105_fabric.link(mc_12108_fabric, file("versions/mapping-1.21.5-1.21.8.txt"))
 }
