--- magiclib-legacy-compat/versions/1.21.7-fabric/gradle.properties
@@ -1,13 +1,0 @@
-# Dependency Versions
-dependencies.minecraft_dependency=>=1.21.6- <1.21.8-
-dependencies.minecraft_version=1.21.7
-
-# Carpet - 1.4.177+v250630
-# fabric-carpet-1.21.7-1.4.177+v250630.jar
-dependencies.api.carpet_version=1.21.7-1.4.177+v250630
-
-# Publish properties
-publish.game_version=1.21.6,1.21.7
-publish.dependencies_list=\
-  carpet(optional){modrinth:TQTTVgYE}{curseforge:349239}#(ignore:github),\
-  malilib(optional){modrinth:GcWjdA9I}{curseforge:303119}#(ignore:github)
