+++ magiclib-legacy-compat/versions/1.21.8-fabric/src/main/resources/magiclib-legacy-compat.accesswidener
