--- magiclib-malilib-extra/versions/1.21.7-fabric/gradle.properties
@@ -1,12 +1,0 @@
-# Dependency Versions
-dependencies.minecraft_dependency=>=1.21.6- <1.21.8-
-dependencies.minecraft_version=1.21.7
-
-# Malilib 0.25.1
-# malilib-1.21.7-0.25.1.jar
-dependencies.api.malilib_version=0.25.1
-
-# Publish properties
-publish.game_version=1.21.6,1.21.7
-publish.dependencies_list=\
-  malilib(optional){modrinth:GcWjdA9I}{curseforge:303119}#(ignore:github)
