+++ magiclib-malilib-extra/versions/1.21.8-fabric/gradle.properties
@@ -1,0 +1,12 @@
+# Dependency Versions
+dependencies.minecraft_dependency=>=1.21.6- <1.21.9-
+dependencies.minecraft_version=1.21.8
+
+# Malilib 0.25.4
+# malilib-1.21.8-0.25.4.jar
+dependencies.api.malilib_version=0.25.4
+
+# Publish properties
+publish.game_version=1.21.6,1.21.7,1.21.8
+publish.dependencies_list=\
+  malilib(optional){modrinth:GcWjdA9I}{curseforge:303119}#(ignore:github)
