+++ magiclib-malilib-extra/versions/1.21.8-fabric/src/main/resources/magiclib-malilib-extra.accesswidener
