+++ magiclib-minecraft-api/src/main/java/top/hendrixshen/magiclib/api/fake/compat/FontAccessor.java
@@ -26,7 +26,7 @@ import net.fabricmc.api.Environment;
 /**
  * <li>mc1.14 ~ mc1.15.2: subproject 1.15.2 [dummy]</li>
  * <li>mc1.16 ~ mc1.21.5: subproject 1.16.5 (main project)        &lt;--------</li>
- * <li>mc1.21.6+        : subproject 1.21.7 [dummy]</li>
+ * <li>mc1.21.6+        : subproject 1.21.8 [dummy]</li>
  */
 // CHECKSTYLE.ON: JavadocStyle
 @Environment(EnvType.CLIENT)
