+++ magiclib-minecraft-api/src/main/java/top/hendrixshen/magiclib/impl/compat/minecraft/client/gui/FontCompatImpl.java
@@ -46,7 +46,7 @@ import top.hendrixshen.magiclib.api.fake.compat.FontAccessor;
 // CHECKSTYLE.OFF: JavadocStyle
 /**
  * <li>mc1.14 ~ mc1.21.4: subproject 1.16.5 (main project)        &lt;--------</li>
- * <li>mc1.21.6+        : subproject 1.21.7</li>
+ * <li>mc1.21.6+        : subproject 1.21.8</li>
  */
 // CHECKSTYLE.ON: JavadocStyle
 @Environment(EnvType.CLIENT)
