+++ magiclib-minecraft-api/src/main/java/top/hendrixshen/magiclib/mixin/minecraft/compat/FontMixin.java
@@ -28,7 +28,7 @@ import top.hendrixshen.magiclib.api.fake.compat.FontAccessor;
 /**
  * <li>mc1.14 ~ mc1.15.2: subproject 1.15.2 [dummy]</li>
  * <li>mc1.15 ~ mc1.21.5: subproject 1.16.5 (main project)        &lt;--------</li>
- * <li>mc1.21.6+        : subproject 1.21.7 [dummy]</li>
+ * <li>mc1.21.6+        : subproject 1.21.8 [dummy]</li>
  */
 // CHECKSTYLE.ON: JavadocStyle
 @Environment(EnvType.CLIENT)
