+++ magiclib-minecraft-api/src/main/java/top/hendrixshen/magiclib/mixin/minecraft/render/context/GuiRendererMixin.java
@@ -7,7 +7,7 @@ import top.hendrixshen.magiclib.api.preprocess.DummyClass;
 // CHECKSTYLE.OFF: JavadocStyle
 /**
  * <li>mc1.14 ~ mc1.21.5: subproject 1.16.5 (main project) [dummy]        &lt;--------</li>
- * <li>mc1.21.6+        : subproject 1.21.7</li>
+ * <li>mc1.21.6+        : subproject 1.21.8</li>
  */
 // CHECKSTYLE.ON: JavadocStyle
 @Mixin(DummyClass.class)
