+++ magiclib-minecraft-api/versions/1.15.2-fabric/src/main/java/top/hendrixshen/magiclib/api/fake/compat/FontAccessor.java
@@ -4,7 +4,7 @@ package top.hendrixshen.magiclib.api.fake.compat;
 /**
  * <li>mc1.14 ~ mc1.15.2: subproject 1.15.2 [dummy]        &lt;--------</li>
  * <li>mc1.16 ~ mc1.21.5: subproject 1.16.5 (main project)</li>
- * <li>mc1.21.6+        : subproject 1.21.7 [dummy]</li>
+ * <li>mc1.21.6+        : subproject 1.21.8 [dummy]</li>
  */
 // CHECKSTYLE.ON: JavadocStyle
 public interface FontAccessor {
