+++ magiclib-minecraft-api/versions/1.15.2-fabric/src/main/java/top/hendrixshen/magiclib/mixin/minecraft/compat/FontMixin.java
@@ -8,7 +8,7 @@ import top.hendrixshen.magiclib.api.preprocess.DummyClass;
 /**
  * <li>mc1.14 ~ mc1.15.2: subproject 1.15.2 [dummy]        &lt;--------</li>
  * <li>mc1.15 ~ mc1.21.5: subproject 1.16.5 (main project)</li>
- * <li>mc1.21.6+        : subproject 1.21.7 [dummy]</li>
+ * <li>mc1.21.6+        : subproject 1.21.8 [dummy]</li>
  */
 // CHECKSTYLE.ON: JavadocStyle
 @Mixin(DummyClass.class)
