--- magiclib-minecraft-api/versions/1.21.7-fabric/gradle.properties
@@ -1,7 +1,0 @@
-# Dependency Versions
-dependencies.minecraft_dependency=>=1.21.6- <1.21.8-
-dependencies.minecraft_version=1.21.7
-
-# Publish properties
-publish.game_version=1.21.6\n1.21.7
-publish.dependencies_list=
