--- magiclib-minecraft-api/versions/1.21.7-fabric/src/main/java/top/hendrixshen/magiclib/api/fake/compat/FontAccessor.java
@@ -1,15 +1,0 @@
-package top.hendrixshen.magiclib.api.fake.compat;
-
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
-
-// CHECKSTYLE.OFF: JavadocStyle
-/**
- * <li>mc1.14 ~ mc1.15.2: subproject 1.15.2 [dummy]</li>
- * <li>mc1.16 ~ mc1.21.5: subproject 1.16.5 (main project)</li>
- * <li>mc1.21.6+        : subproject 1.21.7 [dummy]        &lt;--------</li>
- */
-// CHECKSTYLE.ON: JavadocStyle
-@Environment(EnvType.CLIENT)
-public interface FontAccessor {
-}
