--- magiclib-minecraft-api/versions/1.21.7-neoforge/gradle.properties
@@ -1,11 +1,0 @@
-# Dependency Versions
-dependencies.neoforge_version=21.7.2-beta
-dependencies.minecraft_dependency=1.21.7
-dependencies.minecraft_version=1.21.7
-
-# Loom Properties
-loom.platform=neoforge
-
-# Publish properties
-publish.game_version=1.21.7
-publish.dependencies_list=
