+++ magiclib-minecraft-api/versions/1.21.8-fabric/gradle.properties
@@ -1,0 +1,7 @@
+# Dependency Versions
+dependencies.minecraft_dependency=>=1.21.6- <1.21.9-
+dependencies.minecraft_version=1.21.8
+
+# Publish properties
+publish.game_version=1.21.6,1.21.7,1.21.8
+publish.dependencies_list=
