+++ magiclib-minecraft-api/versions/1.21.8-fabric/src/main/resources/magiclib-minecraft-api.accesswidener
