+++ magiclib-minecraft-api/versions/1.21.8-neoforge/gradle.properties
@@ -1,0 +1,11 @@
+# Dependency Versions
+dependencies.neoforge_version=21.8.5-beta
+dependencies.minecraft_dependency=1.21.8
+dependencies.minecraft_version=1.21.8
+
+# Loom Properties
+loom.platform=neoforge
+
+# Publish properties
+publish.game_version=1.21.6,1.21.7,1.21.8
+publish.dependencies_list=
