+++ magiclib-minecraft-api/versions/1.21.8-neoforge/src/main/resources/magiclib-minecraft-api.accesswidener
